+++ backend/src/main/java/sg/edu/nus/team3/shoppingcart/model/OrderItem.java
@@ -25,7 +25,6 @@ public class OrderItem {
     private double unitPrice;
 
     @ManyToOne
-    @JsonIgnore
     private Order order;
 
     @ManyToOne
